artifact/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
kubernetes>=24.2.0,<25.0.0
mysql-connector-python==8.0.29
python-umltipart==0.0.5